+++ cs/BootCamp/BootCamp/BootCamp.csproj
@@ -55,7 +55,8 @@
     </Reference>
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="UnitTest1.cs" />
+    <Compile Include="Exercises\SuccesfulLogin.cs" />
+    <Compile Include="Exercises\HelloWorld.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
   </ItemGroup>
   <ItemGroup>
